// klifurmot-frontend/src/types/api.ts
export interface ApiSuccessResponse<T> {
    success: true;
    message: string;
    data: T;
    timestamp: string;
}

export interface ApiErrorResponse {
    success: false;
    error: {
        code: string;
        message: string;
        details: Record<string, string[]> | string[];
    };
    timestamp: string;
}